+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN46CKVB5TNXZ9BRFXG1Y5NG"
+        "01KN4PGYQ638A3CNMWD6GWABXS"
     )
     r"""The unique identifier of the data source"""