+++ packages/orq-rc/RELEASES.md
@@ -6779,3 +6779,13 @@ Based on:
 - [python v4.7.0-rc.38] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.38] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.38 - packages/orq-rc
+
+## 2026-04-01 14:21:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.39] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.39] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.39 - packages/orq-rc